+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Improvements
 
 * Improved store reset. `store.reset()` will now reset the store to its initial data, whether or not history is enabled.
+* Add `tableHeader` prop to `Table` component that allows you to set a row wrapped in a `thead` tag.
 
 ## New Components
 
@@ -13,8 +14,11 @@
 
 * Row margin has been reduced to `15px`.
 * Pod component now receives two additional props:
+
   * `border` - allows developers to disable border.
   * `padding` - allows developers to have control over padding size.
+
+* Pill style has changed
 
 ## Improved Dialog